session 1 question 1
Vinoth's model practical is approaching and he hasn't begun to prepare. In order to have the best chance of passing the course, he resolve to study from until practical time. Sections vary in length, but not in value towards a passing mark, so he wants to study as many complete sections as possible. The order of sections doesn't matter, but he must complete a section before it will help Vinoth's mark.

Vinoth's work is to maximize the number of complete sections he can study between now and model practical time.